removed useColorSliderState and moved it to useColorSlider.ts
