+++ packages/react-components/react-color-picker-preview/library/src/components/ColorSlider/useColorSlider.ts
@@ -1,7 +1,19 @@
 import * as React from 'react';
-import { getPartitionedNativeProps, useId, slot } from '@fluentui/react-utilities';
+import { tinycolor } from '@ctrl/tinycolor';
+import {
+  getPartitionedNativeProps,
+  useId,
+  slot,
+  clamp,
+  useControllableState,
+  useEventCallback,
+} from '@fluentui/react-utilities';
+import { colorSliderCSSVars } from './useColorSliderStyles.styles';
+import { useFluent_unstable as useFluent } from '@fluentui/react-shared-contexts';
 import type { ColorSliderProps, ColorSliderState } from './ColorSlider.types';
-import { useColorSliderState_unstable } from './useColorSliderState';
+import { useColorPickerContextValue_unstable } from '../../contexts/colorPicker';
+import { MIN, HUE_MAX as MAX } from '../../utils/constants';
+import { getPercent } from '../../utils/getPercent';
 
 /**
  * Create the state required to render ColorSlider.
@@ -16,6 +28,11 @@ export const useColorSlider_unstable = (
   props: ColorSliderProps,
   ref: React.Ref<HTMLInputElement>,
 ): ColorSliderState => {
+  'use no memo';
+
+  const { dir } = useFluent();
+  const onChangeFromContext = useColorPickerContextValue_unstable(ctx => ctx.requestChange);
+  const colorFromContext = useColorPickerContextValue_unstable(ctx => ctx.color);
   const nativeProps = getPartitionedNativeProps({
     props,
     primarySlotTagName: 'input',
@@ -23,6 +40,8 @@ export const useColorSlider_unstable = (
   });
 
   const {
+    color,
+    onChange = onChangeFromContext,
     vertical,
     // Slots
     root,
@@ -30,6 +49,35 @@ export const useColorSlider_unstable = (
     rail,
     thumb,
   } = props;
+
+  const _color = colorFromContext || color;
+  const hsvColor = tinycolor(_color).toHsv();
+
+  const [currentValue, setCurrentValue] = useControllableState({
+    state: hsvColor.h,
+    initialState: 0,
+  });
+  const clampedValue = clamp(currentValue, MIN, MAX);
+  const valuePercent = getPercent(clampedValue, MIN, MAX);
+
+  const inputOnChange = input?.onChange;
+
+  const _onChange: React.ChangeEventHandler<HTMLInputElement> = useEventCallback(event => {
+    const newValue = Number(event.target.value);
+    const newColor = tinycolor({ ...hsvColor, h: newValue }).toRgbString();
+    setCurrentValue(clamp(newValue, MIN, MAX));
+    inputOnChange?.(event);
+    onChange?.(event, { type: 'change', event, color: newColor });
+    onChangeFromContext(event, {
+      color: newColor,
+    });
+  });
+
+  const rootVariables = {
+    [colorSliderCSSVars.sliderDirectionVar]: vertical ? '180deg' : dir === 'ltr' ? '-90deg' : '90deg',
+    [colorSliderCSSVars.sliderProgressVar]: `${valuePercent}%`,
+    [colorSliderCSSVars.thumbColorVar]: `hsl(${clampedValue}, 100%, 50%)`,
+  };
 
   const state: ColorSliderState = {
     vertical,
@@ -47,6 +95,8 @@ export const useColorSlider_unstable = (
       defaultProps: {
         id: useId('slider-', props.id),
         ref,
+        min: MIN,
+        max: MAX,
         ...nativeProps.primary,
         type: 'range',
       },
@@ -56,7 +106,14 @@ export const useColorSlider_unstable = (
     thumb: slot.always(thumb, { elementType: 'div' }),
   };
 
-  useColorSliderState_unstable(state, props);
+  // Root props
+  state.root.style = {
+    ...rootVariables,
+    ...state.root.style,
+  };
 
+  // Input Props
+  state.input.value = clampedValue;
+  state.input.onChange = _onChange;
   return state;
 };
